+++ packages/spry/example/main.dart
@@ -3,24 +3,7 @@ import 'package:spry/spry.dart';
 final app = Application.late();
 
 void main(List<String> args) async {
-  app.get('/', (request) => 'Hello, Spry!');
-  app.get('/hello/:name', (request) => 'Hello, ${request.params.get('name')}!');
-  app.get('/hello/:name/:age', (request) {
-    final name = request.params.get('name');
-    final age = request.params.get('age');
-
-    return 'Hello, $name! You are $age years old.';
-  });
-  app.get(
-    '/hello',
-    (request) => {
-      '/': 'Hello, Spry!',
-      '/hello': 'Index of hello.',
-      '/hello/:name': 'Say hello to someone.',
-      '/hello/:name/:age': 'Say hello to someone with age.',
-    },
-  );
-  app.post('/hello', (request) async => 'Hello, ${await request.text()}!');
+  app.get('hello', (request) => 'Hello, Spry!');
 
   await app.run(port: 3000);