+++ packages/spry/lib/src/response/responsible.dart
@@ -1,5 +1,5 @@
-import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:path/path.dart' show basename;
 import 'package:webfetch/webfetch.dart' as webfetch;
@@ -7,6 +7,10 @@ import 'package:webfetch/webfetch.dart' as webfetch;
 // ignore: implementation_imports
 import 'package:webfetch/src/_internal/generate_boundary.dart'
     show generateBoundary;
+
+import '../application+encoding.dart';
+import '../request/request+application.dart';
+import '../request/request+json.dart';
 
 abstract interface class Responsible {
   /// Responds to the [request].
@@ -30,16 +34,61 @@ abstract interface class Responsible {
   /// - Dart scalar types, Eg: `String`, `int`, `double`, `bool`...
   factory Responsible.create(Object? object) {
     return switch (object) {
+      webfetch.Blob blob => _StreamResponsible(blob.stream()),
+      webfetch.FormData formData => _FormDataResponsible(formData),
+      webfetch.Response response => _WebfetchResponseResponsible(response),
       Responsible responsible => responsible,
+      HttpClientResponse response => _HttpClientResponsesible(response),
+      HttpResponse _ => const _NullResponsible(),
       Map map => _JsonResponsible(map),
       Iterable iterable => _JsonResponsible(iterable.toList(growable: false)),
       Stream<List<int>> stream => _StreamResponsible(stream),
-      webfetch.FormData formData => _FormDataResponsible(formData),
-      webfetch.Response response => _WebfetchResponseResponsible(response),
+      TypedData typedData => _TypedDataResponsible(typedData),
       File file => _FileResponsible(file),
       null => const _NullResponsible(),
       _ => _NominalResponsible(object),
     };
+  }
+}
+
+class _TypedDataResponsible implements Responsible {
+  final TypedData typedData;
+
+  const _TypedDataResponsible(this.typedData);
+
+  @override
+  Future<void> respond(HttpRequest request) async {
+    final response = request.response;
+    final encoding = request.application.encoding;
+    final bytes = typedData.buffer.asUint8List();
+
+    response.headers.contentType = response.headers.contentType =
+        ContentType('application', 'octet-stream', charset: encoding.name);
+    response.add(bytes);
+  }
+}
+
+class _HttpClientResponsesible implements Responsible {
+  final HttpClientResponse response;
+
+  const _HttpClientResponsesible(this.response);
+
+  @override
+  Future<void> respond(HttpRequest request) async {
+    final innerResponse = request.response;
+
+    innerResponse.statusCode = response.statusCode;
+    innerResponse.contentLength = response.contentLength;
+    innerResponse.cookies.addAll(response.cookies);
+    response.headers.forEach((name, values) {
+      for (final value in values) {
+        innerResponse.headers.add(name, value);
+      }
+    });
+
+    await for (final chunk in response) {
+      innerResponse.add(chunk);
+    }
   }
 }
 
@@ -113,10 +162,14 @@ class _FormDataResponsible implements Responsible {
     final boundary = this.boundary;
 
     final stream = webfetch.FormData.encode(formData, boundary);
-    response.headers.contentType =
-        ContentType('multipart', 'form-data', parameters: {
-      'boundary': boundary,
-    });
+    response.headers.contentType = ContentType(
+      'multipart',
+      'form-data',
+      parameters: {
+        'boundary': boundary,
+      },
+      charset: request.application.encoding.name,
+    );
 
     await response.addStream(stream);
   }
@@ -146,8 +199,12 @@ class _JsonResponsible implements Responsible {
   @override
   Future<void> respond(HttpRequest request) async {
     final response = request.response;
+    final encoding = request.application.encoding;
+    final value = request.jsonCodec.encode(object);
+    final bytes = encoding.encode(value);
 
-    response.headers.contentType = ContentType.json;
-    response.write(json.encode(object));
+    response.headers.contentType = response.headers.contentType =
+        ContentType('application', 'json', charset: encoding.name);
+    response.add(bytes);
   }
 }